Update ioctl platform.machine handling
As the comment in the code says... the lists of possible matches for
UTS_MACHINE name, which can be seen as output of uname -m and which is
the result of doing platform.machine() in python is a bit clumsy.

It will be like this until I learn how to make C extensions, but even in
that case it will keep having this as a fallback to enable users of the
most common archs to just git clone the code and run it, without having
to worry about annoying extra steps to be able to use it.

Kilobyte reported that aarch64 was missing from the list. It's a newer
alias for arm64: https://bugs.debian.org/868370

The bug report caused me to have another look at the UTS_MACHINE
definitions in the linux kernel source, and also simplify a few other
names by doing more startswith matching instead of exact matches.