Rename example show_inode.py -> show_file_by_inode.py
This example code snippet is called show inode, but it can only show
information about a regular file. So, instead of just crashing in a way
like "Exception: Whoa, key DIR_ITEM" if you point it at a directory,
give a bit more meaningful error message.

Also, rename the script to make the purpose more clear.

It's left as an exercise to the reader to create a slightly modified
version of show_directory that works with an inode number of a directory
instead of the name. ;-] After all, the purpose of all the example code
is to help others getting up to speed with doing whatever they want to
do being able to dig around in the live metadata information that's
available in a running Linux kernel.

Reported by Andrei Borzenkov, thanks!