crc32c: Improve crc32c implementation
Add a new function crc32c_data which calculates the corect checksum of a
bytes/bytesarray being passed to it. It uses the core crc32c implementation
and implements the equivalent of btrfs_csum_final e.g. reversing the bits
of the final checksum

Also add an explicit truncation to 32bit value of the seed passed to crc32c
function. This allows to ~1 and ~0 as seed values rather than manually calculating
their 32bit equivalents.