Add free space fragmentation score example
The example code looks at the free space information of all data block
groups and assigns a free space fragmentation score based on the amount
and size of free space extents.

The algorithm used is documented in the script. It's a bit arbitrary,
but after trying a few other things, I found out this approach worked
best for my own situation.

My experience learns that scores above 250 point at chunks in which free
space fragmentation is getting quite bad.

First we try to use the free space tree, because it's the fastest way to
inspect free space. If it's not available, because the filesystem still
uses the old v1 space cache, we can also walk the extent tree (which is
notably more work!) and take the inverse of what we find.