Add example how to read data checksums
Ok, this is a fun one. It shows how to find and unpack checksum item
data from the checksum tree, grab a specific checksum value and then
compare the value with a computed checksum after reading some data from
the example file itself.

However, there's a big disclaimer on top of the output, because this
kind of code should never appear in any serious program. :D

The output of the example tells a whole story about what it's doing.