Implement setting received uuid ioctl
When using send/receive, the received subvolume gets some attributes set
right before it's made read only: received_uuid, rtime, rtransid, stime
and stransid. Because receive is user space code, it uses the ioctl to
do so.

* The received_uuid is set to the uuid of the subvolume that was sent.
* rtime and rtransid is set by the kernel code to current time and
  generation of the fs.
* stime and stransid are provided by us. The receive code puts the
  ctransid value of the sent subvolume into rtransid, and as it seems
  never sets the value of stime, so it's always zero.

Having this ioctl available also means we can abuse it ourselves to play
around with it. :]

Here's the struct from the kernel source:

struct btrfs_ioctl_received_subvol_args {
        char    uuid[BTRFS_UUID_SIZE];  /* in */
        __u64   stransid;               /* in */
        __u64   rtransid;               /* out */
        struct btrfs_ioctl_timespec stime; /* in */
        struct btrfs_ioctl_timespec rtime; /* out */
        __u64   flags;                  /* in */
        __u64   reserved[16];           /* in */
};

The flags are not used at all. The ioctl code also does not do strict
checks to make sure that unused fields in the input are all zeroed out.
This means that the flags and reserved space can't be used for anything
in the future without creating an IOC_SET_RECEIVED_SUBVOL_V2... D:

The input to set_received_subvol is:
* an open file descriptor to the subvolume root directory (inode 256)
* the uuid (a python uuid object) we want to have set as received_uuid
* stransid as a number
* stime we want to have set, as a btrfs.ctree.TimeSpec object

Since all objects in the btrfs.ctree module are created from search
ioctl results, there is no nice way to just create new ones from
scratch. For TimeSpec, I started adding this possibility by adding a
factory method called from_values, which initialized a new object based
on the values we want to have for all attributes.

I yolo'ed around a bit with the args structs to keep the amount of lines
of code down.

Also, an example script is added, which shows how to call this new
function.