Update example for logical to inode v2 ioctl
Now we have the LOGICAL_INO_V2 available, we can use it to overcome the
limitations of the output for this example.

During the first call to the ioctl, we find out if v2 is available. If
so, we keep using it, otherwise we fall back to v1.

We still use the default 64k buffer. But, if v2 is used and we see
missed_bytes, we retry with a buffer that will fit all results.