Register attrs derived from the item key
Some of the attributes we have on objects are created for convenience
reasons. It's much nicer to e.g. access blockgroup.vaddr instead of
having to use blockgroup.key.offset all the time.

But, technically seen these attributes are not part of the corresponding
struct we have in the C code.

The upcoming pretty printer will use this information to show the
difference between values which are present in the item data, and values
which are taken from the item key.

Behaviour of the code for users of the objects does not change.