ExtentItem, MetaDataItem: load refs by default
Commit 6bbeac53 introduced the load_*_refs flags for the ExtentItem and
MetaDataItem objects and set them False by default. The reason for this
is that they can be numerous, and they are mostly meaningless in user
space.

When we directly fill the objects from item data retrieved from the
search ioctl, do load them. This allows us to dynamically load the
header, data tuple into objects where the class is dynamically chosen
based on the object type, without having to code exceptions for specific
item types.

When retrieving information using the fs.extents() helper in the
FileSystem class, it will still be disabled by default.