examples/show_free_space_fragmentation: default min_score 250
The reason to write this example was to determine how bad the result was
of running a filesystem with the <4.14 -o ssd for a long time.

By default, it now outputs information about block groups that have
really badly fragmented free space.

It's possible to specify the minimal score to display as a command line
parameter.

Thanks Kobaz for the feedback!