Add the fideduperange ioctl
This commit implements calling the deduplication ioctl, which started
out as a btrfs specific ioctl once (FILE_EXTENT_SAME), but was moved to
the generic filesystem code and renamed FIDEDUPERANGE. The same number
(based on the 0x94 btrfs ioctl number) is still used, but others can
implement it now. An example is XFS.

When using deduplication, the hard part of the story is not really how
to actually tell btrfs to deduplicate a range of bytes in some files.
The hard part is to actually find out where similar parts of files could
be, in order to have some ranges to feed to the ioctl.

An example program is also added. But, in order to use that example, you
already have to know which files have identical parts in them.

Here's an example how to use the example:

    dd if=/dev/urandom of=src bs=1MiB count=1
    dd if=src of=dst1 bs=1MiB count=1
    dd if=src of=dst2 seek=1 bs=1MiB count=1

    sync

    ./show_file.py src
    ./show_file.py dst1
    ./show_file.py dst2

Right now, you should see that while the 1MiB of data in the files is
the same, every file has its own extents with data.

Now, let's try to free up 2MiB of disk space by having dst1 and dst2
reflink to the extent in src instead:

    ./file_dedupe_range.py src 0 1048576 dst1 0 dst2 1048576

The show_file.py output should now show you that dst1 and dst2 refer to
the same data extent that is used by the file src.

The first version of a patch for dedupe support was submitted by
darkbringer via IRC. Thanks!