Add comment about balance struct names
The btrfs_balance_args and btrfs_balance_progress are actual structs
that exist in the kernel source, they're not just helpers to get the
dirty work done in here. So, they shouldn't be prefixed with an
underscore.

However, we also have a function with the name balance_progress in
here. Argh! That's what you get when stripping off prefixes and moving
part of the name of C structs into the python lib hierarchy...