Introduce ItemNotFoundError
The handling of fs.block_group(vaddr) was a bit stupid actually, because
if no block group is found at that vaddr, the return block_groups[0]
explodes with the IndexError because the list is empty...

With a more specific error, the calling code wouldn't accidentally also
catch an IndexError because of a bug in other code, like the search_v2
which it's using.

We have to inherit the ItemNotFoundError from IndexError to prevent
breaking checks in existing code that uses this.