Work around circular imports
Instead of doing second level imports, just import btrfs at the
beginning always.

The trick is to not use things before they get imported. An import
btrfs.ctree can fail because ctree is not in there, because the current
module gets imported because of an earlier import in btrfs __init__.py
etc...