Add volumes.py, handling device / chunk logic
In the kernel, this is in volumes.c, so keep it the same.

The kernel has an ugly array with attributes per raid type, called
raid_attrs, or raid_array. Let's copy it in here.  We need these
properties to build a few functions which can calculate interesting
things about chunks and block groups.

I added a new field, nparity, which is needed information to make this
happen.

Right now, there's five of them already:

1. chunk_to_dev_extent_length(chunk) and
2. chunk_length_to_dev_extent_length(flags, num_stripes, chunk_length)

Given a chunk metadata object, it calculates what the length of the
device extent is (they're all the same size). The second function does
the same, when providing the needed data explicitely.

3. dev_extent_length_to_chunk_length(flags, num_stripes, stripe_size)

This is the inverse of the first two functions.

4. chunk_to_raw_parity_bytes(chunk)

Given a chunk metadata object, this calculates how many bytes are
reserved for storing parity data for RAID56 profiles. The number is
relevant to understand how much raw disk space that is allocated but not
used is actually not usable for data, because it's reserved for parity.

5. block_group_profile_raw_data_ratio(flags)

This function returns how many times the actual data is replicated on
disk given block group flags as input. E.g. for RAID1 this is 2, but for
RAID5, this is just 1. The actual data is stored once in case of RAID5,
and the redundancy is done using parity blocks, not data itself.