FsUsage, a detailed usage calculator
--- >8 ---

Background

The utils module currently has some functionality to determine space
usage and disk usage, which is used by the munin and nagios monitoring
plugin. That code is one of the first things I wrote for this library.
However, it suffers from being very naive, and not supporting the RAID56
types.

I had another look at this, and for RAID56 this means getting an idea
about how much allocated space is reserved for parity. For total
numbers, there's a quick way to do this, since on device level,
allocated usage included parity, and totals of 'spaces' do not include
the parity bytes. So, taking the difference and assuming it's the bytes
reserved for parity is enough.

--- >8 ---

But, I wanted to have a bit more detailed report of usage, since there's
quite some scenarios in which this is useful.

When creating an FsUsage object with btrfs.utils.FsUsage(fs), the
resulting object will contain:

FsUsage - Total, unallocated, allocated, parity and used disk space for
 |        the btrfs physical address space. Total and used bytes for the
 |        virtual address space.
 \_ DevUsage - Total and allocated physical bytes per device.
 |  \_ DevSpaceUsage - Allocated and parity bytes per space per device.
 \_ RawSpaceUsage - Allocated, parity and used raw bytes per space.
 \_ BlockGroupTypeUsage - Allocated, parity and used bytes per block
 |                        group type (system, data, metadata)
 \_ VirtualSpaceUsage - Total and used virtual bytes per space.
 \_ VirtualBlockGroupTypeUsage - Total and used virtual bytes per block
                                 group type (system, data, metadata)

With this info, it's possible to construct output like the btrfs dev
usage and btrfs fi usage show, and many more overviews. But, it's also
possible to use it for e.g. creating a smart balance algorithm that does
the minimal amount of work to get the filesystem balanced after changing
disks.

Hint: try creating a FsUsage object and then use
btrfs.utils.pretty_print to show it:

    import btrfs
    with btrfs.FileSystem('/') as fs:
        btrfs.utils.pretty_print(fs.usage())

Since the amount of code involved is a bit more than I anticipated, I
put it in its own new module, fs_usage, instead of in utils.

--- >8 ---

About the calculation of unallocatable space:

By using a slightly simplified copy of the chunk allocator logic from
the kernel, we can just simulate allocating all available raw disk
space, and see whatever remains as left-over unallocatable space.

The result is two sets of information:
1. 'soft' wasted space: This is the unallocatable space when continuing
   to use the filesystem in its current state, adding new data and
   metadata.
2. 'hard' wasted space: The unallocatable space when the filesystem
   would be optimally balanced.

The difference between the soft and hard wasted space is raw disk space
than can be reclaimed when using btrfs balance to move data around over
the devices in the filesystem.

Note: The results are approximates. When using this, don't complain
about the algorithm reporting 1 or 2 GiB of unallocatable space. Use it
to discover that you have many GiBs or TiBs of unallocatable space.

--- >8 ---

Last but not least... I started adding a lot of documentation, which can
be transformed into nice html pages with sphinx.