__init__: add pretty printing property factory
We have quite a few attributes that we want to have pretty printed when
feeding entire objects to the pretty printer. Let's make it a little bit
less cumbersome to generate the code to do that. Instead of writing out
all the property code we can generate it.

This machinery will add the <foo>_str properties to classes with the
designated function handling the conversion to pretty value.