ctree: fix Key wrap-around
In commit a9a167c26d, I show that subtracting 1 from a zeroed key will
provide the maximum possible key value. However, even the example in the
commit message shows that there's a problem, and apparently I totally
missed that for some obscure reason.

    >>> import btrfs
    >>> key = btrfs.ctree.Key(0, 0, 0)
    >>> key - 1
    Key(-1, 255, 18446744073709551615)
    >>> key._objectid
    0
    >>> max_key = key - 1
    >>> max_key._objectid
    -1

This is not good. This should be 18446744073709551615 as well.

So, add missing ULL() wrapper to make sure we have a positive number.
While doing so, also replace the explicit (1 << 64) - 1) for offset with
ULL() and provide a new U8() to only keep one byte.

Now we get this:

    >>> import btrfs
    >>> key = btrfs.ctree.Key(0, 0, 0)
    >>> key - 1
    Key(18446744073709551615, 255, 18446744073709551615)