Fix retrieving >=2^32 items from a search run
The search ioctl in the kernel has a nr_items field which limits the
amount of items that will be written to the memory buffer we provide
during a single search call.

However, in this library, in the search(v2) wrapper functions, the
nr_items parameter behaves a little bit different, because it's the
max amount of items we want in total, even when multiple search calls
are needed to get all of them. The search helper just generates results,
and it will call the kernel again and again until there's no result any
more, or until we reached nr_items.

The max nr_items that can be set for a single search call is 2^32-1.
But, the default nr_items here was also set to this value. Don't do
that, since it limits the grand total.

Instead: If nr_items is set, use it as before. If it's not set, then
just use the 2^32-1 maximum for every individual search call.