ctree: key attributes for directory index objects
InodeItem, InodeRefList, InodeExtrefList, DirItemList and DirIndex all
have information in the item key that is relevant. Mostly it's inode
numbers for the current file and containing directory.

Let's put them in attributes, so they will show up when using the pretty
printer. Also show them in str() output.

An InodeItem can also be embedded in a RootItem, in which case it
doesn't have its own key. So, we need to handle also that scenario with
some ifery.