ctree: add INODE_ROOT_ITEM_INIT
The RootItem has an embedded InodeItem, and in there, inside the flags
field, a flag is reserved for the RootItem. If the flag is set, it means
that flags and byte_limit fields of the RootItem are properly
initialized for subvolume trees. This was added in Linux 2.6.39.

It's there, so show it properly when pretty printing the flags.