ctree,utils: create Key from string representation
So, we have all kinds of code to pretty print a metadata key. Now, next
step is to add the reverse, and be able to parse key text back to a Key
object.

When initializing a Key object, it's now also allowed to use text
representations of the numeric values, like using 'DEV_EXTENT' instead
of 204 as type value.

To parse entire pretty formatted text strings, there's a helper function
in utils, parse_key_string, which can handle them, e.g. '(31832
INODE_REF 31798)' as input, and do whatever is necessary to dissect the
text and return a Key object.