btrfs-search-metadata: Hello, world!
After moving a bunch of interesting example programs into the bin/
location I'm looking at the rest of the random pile of code in examples/

What about a tool that allows searching any metadata you want? And what
if it already has a bunch of presets that select ranges from trees
automatically?

    btrfs-search-metadata [--format {keys,short,long}] <preset> [<moar> ..] <path>

The first version of this program has preset search options for:
- chunks
- block groups
- device extents
- all information for an inode by pointing it to a file or directory
- all information for an inode (objectid) in any subvolume
- devices (the metadata items)
- orhpans (the orphan items from the root tree)
- dump: custom search query for a range of items in a tree
- block_group_contents: show block group contents
- block_group_free_space: decoded free space extents inside a block group

This tool will be used in the tutorial a lot, when looking around and
learning about all what's to see and how it relates together.