ctree: fix key value wrap behaviour
We can do fun stuff with keys, like this:

    >>> k = btrfs.ctree.Key(0, 0, 0) - 1
    >>> k
    Key(18446744073709551615, 255, 18446744073709551615)

However... what's the actual 136 bit key value in this case?

    >>> k.key
    -1

That's not good, because -1 is not a valid unsigned 136 bit number at
all. That should be:

    >> btrfs.ctree.Key(-1, -1, -1).key
    87112285931760246646623899502532662132735

So, when using  btrfs.ctree.Key(0, 0, 0) - 1 as max key value for a
search, only the values from the first search ioctl call result buffer
would be returned, and then the search would stop.

Same bug is present for wrapping the value the other way round:

    >>> l = btrfs.ctree.Key(18446744073709551615, 255, 18446744073709551615) + 1
    >>> l
    Key(0, 0, 0)
    >>> l.key
    87112285931760246646623899502532662132736

So, we have to manually reduce the value to 136 bits again, since python
has no idea about different sized number storage.

Fixes: 2916bbf54d ("Search ioctl, search key, devices, chunks and stripes")