ctree: add search function to FileSystem
This is just a wrapper around ioctl.search_v2 so that we can write
simple searches with even less verbose code.

Don't bother with situations that need falling back to search v1, since
if you're running a Linux kernel < 3.16 with btrfs, then instead of
playing around with this code, you need to upgrade.