ctree: fix unnecessary multiple root item handling
So, in commit 606a5b5fd4 ("Retrieve subvolume info"), there's a whole
story about multiple root tree items for a single subvolume, because I
found all kinds of documentation about that in the btrfs wiki. However,
it seems this is already outdated information since 2009.

So, simplify the code and forget about all of this, kthxbye.

Fixes: 606a5b5fd4 ("Retrieve subvolume info")
Link: https://lore.kernel.org/linux-btrfs/f0144a34-f14a-5413-9b0c-a2ccba1a1cd9@knorrie.org/