ctree: add OrphanItem
These live in the root tree. One way to spot them is to throw away a
subvolume. The tree id will show up as orphan item while it's not
completely removed in the background.

This item has no data, only an item key. For convenience reasons we put
the objectid in an attribute.