debian/control: use python3:any
For a related bug report, see #968501 for btrfs-heatmap. The same
applies here. The ${python3:Depends} for the actual python3-btrfs
package also results in using python3:any since it's Architecture:
all.

Note that the dependency is not technically needed, since they already
depend on python3-btrfs, but lintian is not smart enough to understand.
If we would leave it out, the missing-dep-for-interpreter lintian error
will show up.