btrfs-balance-least-used: sort block groups by bytes used, not percentage used
This way, we can quickly clean up completely empty block groups.

Additionally, if `btrfs-balance-least-used -u 0` is invoked, only
consider fully empty block groups because this seems a useful edge case
(there can be a lot of almost-empty block groups which are technically
used_pct==0 but still take a lot of time to balance, so the user may as
well invoke with `-u 1`).