ctree: FileSystem: add block_groups function
In the past, I did not really want to add this helper function, since
it's not just translating some function parameters to another function
call, but, to get all Block Group objects, we needed to search the Chunk
tree and get all of them individually. So, to make it more explicit to
the user of the library that it was a bit weird inefficient process, I
let the user do that little dance.

Now, with the new Block Group Tree, we can actually just to a cheap
lookup of a Block Group range! So, well, let's add the convenience
function now, and let it handle both the old and new case.

Note that the difference in behaviour between error handling for looking
up a range or a single items stays the same. block_groups(...) will
return an iterator which has no objects to produce, and block_group(...)
will throw the ItemNotFoundError.