debian/control: drop munin-node dependency
I added this dependency right at the beginning of the
munin-plugins-btrfs already, probably because I was thinking that it
makes no sense to install the plugin without also having munin-node.

However, technically, there is no dependency, and users might also want
to use this plugin with an alternative like munin-node-c.

Reported-by: Marc Donges <marc.donges@gmail.com>
Closes: #1079373