Create a Python btrfs gallery
Let's start a separate page to show programs that are using this
library. This is nice to browse around for new users.

Currently the main examples in python-btrfs itself, and of course
btrfs-heatmap are listed.

Anyone who has created an interesting tool using python-btrfs can now
request to be listed on this page! \o/